requirements.txt
kagglehub
mysql-connector-python
tabulate

